# packages/shared-types/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/posts": {
      "get": {
        "operationId": "PostController_getAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Post"
                  }
                }
              }
            }
          }
        },
        "summary": "Get all posts",
        "tags": [
          "posts"
        ]
      },
      "post": {
        "operationId": "PostController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePostDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a post (school only)",
        "tags": [
          "posts"
        ]
      }
    },
    "/api/posts/school/{schoolId}": {
      "get": {
        "operationId": "PostController_getBySchool",
        "parameters": [
          {
            "name": "schoolId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Post"
                  }
                }
              }
            }
          }
        },
        "summary": "Get posts by school",
        "tags": [
          "posts"
        ]
      }
    },
    "/api/posts/{id}": {
      "get": {
        "operationId": "PostController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              }
            }
          }
        },
        "summary": "Get post by ID",
        "tags": [
          "posts"
        ]
      }
    },
...
    "/api/ratings/post/{postId}": {
      "get": {
        "operationId": "RatingController_getByPost",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Rating"
                  }
                }
              }
            }
          }
        },
        "summary": "Get ratings for a post",
        "tags": [
          "ratings"
        ]
      }
    },
...
  },
  "info": {
    "title": "AirConnect Hanoi API",
    "description": "API for AirConnect Hanoi - Air Quality Management Platform",
    "version": "2.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "Post": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "enum": [
              "case_study",
              "best_practice",
              "research",
              "news",
              "guide",
              "other"
            ],
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "image": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "published_at": {
            "format": "date-time",
            "type": "string"
          },
          "school_id": {
            "type": "number"
          },
          "school": {
            "$ref": "#/components/schemas/School"
          },
          "ratings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Rating"
            }
          }
        },
        "required": [
          "id",
          "title",
          "type",
          "description",
          "image",
          "content",
          "published_at",
          "school_id",
          "school",
          "ratings"
        ]
      },
      "Rating": {
        "type": "object",
        "properties": {
          "post_id": {
            "type": "number"
          },
          "student_id": {
            "type": "number"
          },
          "rate": {
            "enum": [
              "1",
              "2",
              "3",
              "4",
              "5"
            ],
            "type": "string"
          },
          "rated_at": {
            "format": "date-time",
            "type": "string"
          },
          "post": {
            "$ref": "#/components/schemas/Post"
          },
          "student": {
            "$ref": "#/components/schemas/Student"
          }
        },
        "required": [
          "post_id",
          "student_id",
          "rate",
          "rated_at",
          "post",
          "student"
        ]
      },
      "Student": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "first_name": {
            "type": "string"
          },
          "last_name": {
            "type": "string"
          },
          "sex": {
            "enum": [
              "male",
              "female"
            ],
            "type": "string"
          },
          "birth_date": {
            "format": "date-time",
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "health_status": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "school_id": {
            "type": "number"
          },
          "school": {
            "$ref": "#/components/schemas/School"
          },
          "ratings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Rating"
            }
          }
        },
        "required": [
          "id",
          "first_name",
          "last_name",
          "sex",
          "birth_date",
          "email",
          "phone",
          "health_status",
          "password",
          "created_at",
          "school_id",
          "school",
          "ratings"
        ]
      },
      "Admin": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "type": {
            "type": "string",
            "enum": [
              "super",
              "normal"
            ]
          },
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "reviews": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Review"
            }
          }
        },
        "required": [
          "id",
          "type",
          "username",
          "password",
          "phone",
          "email",
          "created_at",
          "reviews"
        ]
      },
      "Review": {
        "type": "object",
        "properties": {
          "submission_id": {
            "type": "number"
          },
          "admin_id": {
            "type": "number"
          },
          "decision": {
            "enum": [
              "pending",
              "accepted",
              "rejected",
              "other"
            ],
            "type": "string"
          },
          "date": {
            "format": "date-time",
            "type": "string"
          },
          "note": {
            "type": "string"
          },
          "admin": {
            "$ref": "#/components/schemas/Admin"
          },
          "submission": {
            "$ref": "#/components/schemas/Submission"
          }
        },
        "required": [
          "submission_id",
          "admin_id",
          "decision",
          "date",
          "note",
          "admin",
          "submission"
        ]
      },
      "Submission": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "type": {
            "enum": [
              "request",
              "offer",
              "other"
            ],
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "from_school_id": {
            "type": "number"
          },
          "school": {
            "$ref": "#/components/schemas/School"
          },
          "reviews": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Review"
            }
          }
        },
        "required": [
          "id",
          "type",
          "content",
          "created_at",
          "from_school_id",
          "school",
          "reviews"
        ]
      },
      "Solution": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "type": {
            "enum": [
              "improving_facilities",
              "research_and_development",
              "other"
            ],
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "note": {
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "status": {
            "enum": [
              "normal",
              "important",
              "critical",
              "other"
            ],
            "type": "string"
          },
          "applies": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Apply"
            }
          }
        },
        "required": [
          "id",
          "type",
          "content",
          "note",
          "created_at",
          "status",
          "applies"
        ]
      },
      "Apply": {
        "type": "object",
        "properties": {
          "solution_id": {
            "type": "number"
          },
          "air_quality_id": {
            "type": "number"
          },
          "applied_at": {
            "format": "date-time",
            "type": "string"
          },
          "solution": {
            "$ref": "#/components/schemas/Solution"
          },
          "air_quality": {
            "$ref": "#/components/schemas/AirQuality"
          }
        },
        "required": [
          "solution_id",
          "air_quality_id",
          "applied_at",
          "solution",
          "air_quality"
        ]
      },
      "AirQuality": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "aqi": {
            "type": "number"
          },
          "pm25": {
            "type": "number"
          },
          "pm10": {
            "type": "number"
          },
          "co2": {
            "type": "number"
          },
          "temp": {
            "type": "number"
          },
          "humidity": {
            "type": "number"
          },
          "wind_speed": {
            "type": "number"
          },
          "measured_at": {
            "format": "date-time",
            "type": "string"
          },
          "school_id": {
            "type": "number"
          },
          "school": {
            "$ref": "#/components/schemas/School"
          },
          "applies": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Apply"
            }
          }
        },
        "required": [
          "id",
          "aqi",
          "pm25",
          "pm10",
          "co2",
          "temp",
          "humidity",
          "wind_speed",
          "measured_at",
          "school_id",
          "school",
          "applies"
        ]
      },
      "Help": {
        "type": "object",
        "properties": {
          "from_school_id": {
            "type": "number"
          },
          "to_school_id": {
            "type": "number"
          },
          "type": {
            "enum": [
              "request",
              "offer",
              "other"
            ],
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "status": {
            "enum": [
              "pending",
              "accepted",
              "rejected",
              "other"
            ],
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "from_school": {
            "$ref": "#/components/schemas/School"
          },
          "to_school": {
            "$ref": "#/components/schemas/School"
          }
        },
        "required": [
          "from_school_id",
          "to_school_id",
          "type",
          "content",
          "status",
          "created_at",
          "from_school",
          "to_school"
        ]
      },
      "School": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "type": {
            "type": "string",
            "enum": [
              "primary",
              "secondary",
              "high_school",
              "university"
            ]
          },
          "name": {
            "type": "string"
          },
          "location": {
            "type": "object"
          },
          "address": {
            "type": "string"
          },
          "district": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "situation": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "score_1": {
            "type": "number"
          },
          "score_2": {
            "type": "number"
          },
          "score_3": {
            "type": "number"
          },
          "score_4": {
            "type": "number"
          },
          "score_5": {
            "type": "number"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "students": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Student"
            }
          },
          "submissions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Submission"
            }
          },
          "posts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Post"
            }
          },
          "air_qualities": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AirQuality"
            }
          },
          "helps_from": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Help"
            }
          },
          "helps_to": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Help"
            }
          }
        },
        "required": [
          "id",
          "type",
          "name",
          "location",
          "address",
          "district",
          "password",
          "situation",
          "email",
          "phone",
          "score_1",
          "score_2",
          "score_3",
          "score_4",
          "score_5",
          "created_at",
          "students",
          "submissions",
          "posts",
          "air_qualities",
          "helps_from",
          "helps_to"
        ]
      },
...
      "CreatePostDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "case_study",
              "best_practice",
              "research",
              "news",
              "guide",
              "other"
            ]
          },
          "description": {
            "type": "string"
          },
          "image": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "school_id": {
            "type": "number"
          },
          "published_at": {
            "type": "string"
          }
        },
        "required": [
          "title",
          "type",
          "description",
          "image",
          "content",
          "school_id"
        ]
      },
...
    }
  }
}